requirements.txt
pandas
numpy
matplotlib
seaborn
plotly
kaleido
fastapi
uvicorn
langchain 
langgraph 
langchain-groq 
python-dotenv